pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-get"
version = "0.1.0"
description = "The package manager for MCP servers — install and configure in one command."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["mcp", "model-context-protocol", "claude", "cursor", "ai", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "typer>=0.12",
    "rich>=13.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
yaml = ["pyyaml>=6.0"]
all  = ["pyyaml>=6.0"]

[project.urls]
Homepage = "https://github.com/shalayiding/mcp-get"
Repository = "https://github.com/shalayiding/mcp-get"
Issues = "https://github.com/shalayiding/mcp-get/issues"

[project.scripts]
mcp = "mcp_hub.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["mcp_hub"]
